+++ model/meditationFees.js
@@ -11,7 +11,7 @@ const meditationFees = sequelize.define('meditationFees', {
     razorpay_payment_id: { type: DataTypes.STRING},
     razorpay_signature: { type: DataTypes.STRING}, 
     UId:{ type: DataTypes.INTEGER}  ,
-    amount : { type : DataTypes.INTEGER},
+    amount : { type : DataTypes.DOUBLE},
     payment_date :{ type:DataTypes.STRING},
     payment_time:{ type: DataTypes.STRING},
     fee_payment_status:{ type: DataTypes.BOOLEAN},
